+++ tedi/components/helpers/ellipsis/index.ts
@@ -1,0 +1,2 @@
+export { EllipsisComponent } from "./ellipsis.component";
+export type { EllipsisPosition } from "./ellipsis.component";